+++ routes/admin.js
@@ -1,0 +1,8 @@
+const router = require('express').Router()
+
+/* GET admin page */
+router.get('/', (req, res) => {
+    res.render('admin.ejs')
+})
+
+module.exports = router
