+++ routes/logout.js
@@ -1,0 +1,8 @@
+const router = require('express').Router()
+
+/* GET logout page */
+router.get('/submit', (req, res) => {
+    req.session = null
+    res.redirect('/login')
+})
+module.exports = router
