+++ routes/root.js
@@ -1,0 +1,12 @@
+const router = require('express').Router()
+
+/* GET root page */
+router.get('/', (req, res) => {
+    if (req.session.user) {
+        res.redirect('/home')
+    } else {
+        res.render('root.ejs')
+    }
+})
+
+module.exports = router
